Added registration of per-directory config (refactor)
Massive refactor to make the actual brancher build tool
a service. All build code was moved to a new Brancher class.

Added a IBrancherExtension interface, which can be registered
with the Brancher service in order to hook into events, receive
configuration from a new per-directory special file functionality,
etc.

Also added a built-in PostGeneratorExtension, which allows
generating files in a directory automatically based on a data
source.