Added live-updating server and template caching
New system where templates keep track of what they depend
on, including other templates and assets. This allows for
tracking of whether a template actually needs to be
re-rendered.

Also added a new live-updating server that uses the template
caching system and possibly inotify to detect whe it needs to
update.

Fixes #13